lroberts comments
