+++ src/interface/update.hpp
@@ -29,6 +29,7 @@
 #include "interface/state_descriptor.hpp"
 #include "pack/make_pack_descriptor.hpp"
 #include "pack/sparse_pack.hpp"
+#include "sparse/sparse_management.hpp"
 #include "time_integration/staged_integrator.hpp"
 
 #include "kokkos_abstraction.hpp"
@@ -312,76 +313,11 @@ TaskStatus FillDerived(T *rc) {
   return TaskStatus::complete;
 }
 
-// TODO(JMM): Should this move into sparse_management?
-template <typename T>
-TaskStatus InitNewlyAllocatedVars(T *rc) {
-  PARTHENON_INSTRUMENT
-  if (!rc->AllVariablesInitialized()) {
-    const IndexDomain interior = IndexDomain::interior;
-    const IndexRange ib = rc->GetBoundsI(interior);
-    const IndexRange jb = rc->GetBoundsJ(interior);
-    const IndexRange kb = rc->GetBoundsK(interior);
-    const int Ni = ib.e + 1 - ib.s;
-    const int Nj = jb.e + 1 - jb.s;
-    const int Nk = kb.e + 1 - kb.s;
-    const int NjNi = Nj * Ni;
-    const int NkNjNi = Nk * NjNi;
-
-    // This pack will always be freshly built, since we only get here if sparse data
-    // was allocated and hasn't been initialized, which in turn implies the cached
-    // pack must be stale.
-    auto desc =
-        parthenon::MakePackDescriptor<variable_names::any>(rc, {Metadata::Sparse});
-    auto v = desc.GetPack(rc);
-
-    Kokkos::parallel_for(
-        PARTHENON_AUTO_LABEL,
-        Kokkos::TeamPolicy<>(parthenon::DevExecSpace(), v.GetNBlocks(), Kokkos::AUTO),
-        KOKKOS_LAMBDA(parthenon::team_mbr_t team_member) {
-          const int b = team_member.league_rank();
-          int lo = v.GetLowerBound(b, variable_names::any());
-          int hi = v.GetUpperBound(b, variable_names::any());
-
-          for (int vidx = lo; vidx <= hi; ++vidx) {
-            if (!v(b, vidx).initialized) {
-              Real val = v(b, vidx).sparse_default_val;
-              Kokkos::parallel_for(Kokkos::TeamThreadRange<>(team_member, NkNjNi),
-                                   [&](const int idx) {
-                                     const int k = kb.s + idx / NjNi;
-                                     const int j = jb.s + (idx % NjNi) / Ni;
-                                     const int i = ib.s + idx % Ni;
-                                     v(b, vidx, k, j, i) = val;
-                                   });
-            }
-          }
-        });
-
-    // Set initialized here since everything has been filled with default values,
-    // user defined functions may overwrite these in the next step but that doesn't
-    // change initialization status of the interior
-    rc->SetAllVariablesToInitialized();
-  }
-
-  // Do user defined initializations if present
-  // This has to be done even in the case where no blocks have been allocated
-  // since the boundaries of allocated blocks could have received default data
-  // In any case
-  auto pm = rc->GetParentPointer();
-  for (const auto &pkg : pm->packages.AllPackages()) {
-    pkg.second->InitNewlyAllocatedVars(rc);
-  }
-
-  // Don't worry about flagging variables as initialized
-  // since they will be flagged at the beginning of the
-  // next step in the evolution driver
-
-  return TaskStatus::complete;
-}
-
-TaskStatus SparseDealloc(MeshData<Real> *md);
+// In sparse/sparse_management.hpp but convenient to include here
+using parthenon::InitNewlyAllocatedVars;
+using parthenon::SparseDealloc;
 
 } // namespace Update
-
 } // namespace parthenon
 
 #endif // INTERFACE_UPDATE_HPP_
